Move conversion methods into fulfil/converter
